Narrow the player velocity through an int
pmove_state_t.velocity is a short, and a trigger_push defaults to 1000
units, which the touch scales by ten and this line by eight -- far outside
the range. Converting an out-of-range double straight to a short is
undefined, and on -m32 GCC emits a 16-bit fistp, whose overflow result is
the x87 indefinite 0x8000. So a jump pad pinned the axis to -4096 with the
sign lost, in both directions.

Going through an int gives the wrap the 1999 build produced, which is what
the client's own prediction expects.

The origin line above needs no cast: at the map limit its wrap and the
indefinite value coincide.

+++ src/p_client.c
@@ -2579,7 +2579,7 @@ chasing:
 	for (i=0 ; i<3 ; i++)
 	{
 		pm.s.origin[i] = ent->s.origin[i]*8;
-		pm.s.velocity[i] = ent->velocity[i]*8;
+		pm.s.velocity[i] = (int)(ent->velocity[i]*8);
 	}
 
 	if (memcmp(&client->old_pmove, &pm.s, sizeof(pm.s)))
